+++ lib/src/lazy.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 /// A class that provides lazy initialization for a value of type [T].
 /// The value is created by a factory function the first time it is needed.
 /// Optionally, the value can expire based on a duration or a custom condition,
@@ -6,11 +8,11 @@ class Lazy<T> {
   T? _instance;
   DateTime? _lastInitializedTime;
 
-  final T Function() _factory;
-  final void Function(T instance)? _onCreate;
+  final FutureOr<T> Function() _factory;
+  final FutureOr<void> Function(T instance)? _onCreate;
   final Duration? _expiryDuration;
-  final bool Function(T instance)? _shouldExpire;
-  final void Function(T instance)? _onDestroy;
+  final FutureOr<bool> Function(T instance)? _shouldExpire;
+  final FutureOr<void> Function(T instance)? _onDestroy;
 
   /// Constructs an instance of `Lazy<T>` with the required factory function
   /// and optional callbacks for creation, expiry, and destruction of the value.
@@ -22,11 +24,11 @@ class Lazy<T> {
   /// - [shouldExpire]: An optional condition that, if returns true, will mark the value as expired and trigger its recreation.
   /// - [onDestroy]: An optional callback that is invoked before the value is reset or recreated, useful for cleanup.
   Lazy({
-    required T Function() factory,
-    void Function(T instance)? onCreate,
+    required FutureOr<T> Function() factory,
+    FutureOr<void> Function(T instance)? onCreate,
     Duration? expiryDuration,
-    bool Function(T instance)? shouldExpire,
-    void Function(T instance)? onDestroy,
+    FutureOr<bool> Function(T instance)? shouldExpire,
+    FutureOr<void> Function(T instance)? onDestroy,
   })  : _factory = factory,
         _onCreate = onCreate,
         _expiryDuration = expiryDuration,
@@ -38,25 +40,26 @@ class Lazy<T> {
 
   /// Returns the current value, creating it if necessary. If the value is determined to be expired
   /// either by time or a custom condition, it is recreated. This check happens each time the value is accessed.
-  T get value {
+  FutureOr<T> get value async {
     final now = DateTime.now().toUtc();
 
     // Check if the value has expired based on time or a custom condition.
     final isTimeBasedExpired = _expiryDuration != null &&
         _lastInitializedTime != null &&
         now.difference(_lastInitializedTime!) > _expiryDuration!;
-    final isExpired =
-        (_instance != null && (_shouldExpire?.call(_instance as T) ?? false)) ||
-            isTimeBasedExpired;
+    final isExpired = (_instance != null &&
+            (await _shouldExpire?.call(_instance as T) ?? false)) ||
+        isTimeBasedExpired;
 
     if (isExpired) {
       reset(); // Reset if expired, ensuring onDestroy is called if provided.
     }
 
     if (_instance == null) {
-      _instance = _factory();
+      _instance = await _factory();
       _lastInitializedTime = now;
-      _onCreate?.call(_instance as T); // Invoke onCreate callback after creation.
+      _onCreate
+          ?.call(_instance as T); // Invoke onCreate callback after creation.
     }
     return _instance!;
   }