+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- The Functions are now returning a FutureOr
+- The value is now an async getter
+
 ## 1.0.0
 
 - Initial version.